Revert "WIP - FSM INBOX (#1032)"
This reverts commit a5b8be200a286eb903553bc63655ca5785b9f801.
