+++ packages/libraries/src/hooks/useInbox.js
@@ -4,7 +4,7 @@ import { InboxGeneral } from "../services/elements/InboxService"
 
 const useInbox = ({tenantId, filters, config}) => useQuery(
         ["INBOX_DATA",tenantId, ...Object.keys(filters)?.map( e => filters?.[e] )],
-        () => InboxGeneral.Search({inbox: {...filters}}),
+        () => InboxGeneral.Search({tenantId, inbox: {...filters}}),
         { ...config }
     )
